Create README.md
This is a sample project that was originally performed as a coding test for and Android developer position. It's an app that consumes the FourSquare API to show a list of businesses within a specific location. The focus of the app is towards architecture and clean code rather than UI/UX. The code should give an indication of the general architecture for a project that could potentially scale to contain dozens features build by a large team of developers. 

Key technologies used in the project:

- Kotlin
- Coroutines
- Compose
- Hilt
- Room
- Retrofit